+++ kyma/training/__init__.py
@@ -1,0 +1,1 @@
+"""Training surfaces for Kyma."""